feat(stm): use circuit witness in prover input
