+++ mithril-stm/src/circuits/halo2/circuit.rs
@@ -16,12 +16,9 @@ use crate::circuits::halo2::errors::{StmCircuitError, to_synthesis_error};
 use crate::circuits::halo2::gadgets::{
     assert_lottery_won, verify_merkle_path, verify_unique_signature,
 };
-use crate::circuits::halo2::types::{
-    CircuitBase, CircuitCurve, CircuitInstance, CircuitMerkleTreeLeaf, MerklePath,
-};
+use crate::circuits::halo2::types::{CircuitBase, CircuitCurve, CircuitInstance, CircuitWitness};
 use crate::signature_scheme::{
     DOMAIN_SEPARATION_TAG_LOTTERY, DOMAIN_SEPARATION_TAG_SIGNATURE, PrimeOrderProjectivePoint,
-    UniqueSchnorrSignature,
 };
 use crate::{LotteryIndex, Parameters, StmResult};
 
@@ -155,12 +152,7 @@ impl StmCircuit {
 
 impl Relation for StmCircuit {
     type Instance = CircuitInstance;
-    type Witness = Vec<(
-        CircuitMerkleTreeLeaf,
-        MerklePath,
-        UniqueSchnorrSignature,
-        LotteryIndex,
-    )>;
+    type Witness = CircuitWitness;
 
     fn format_instance(instance: &Self::Instance) -> Result<Vec<CircuitBase>, Error> {
         Ok(vec![instance.0.into(), instance.1.into()])
