+++ mithril-stm/src/circuits/mod.rs
@@ -10,4 +10,4 @@ pub mod halo2;
 #[cfg(test)]
 pub(crate) mod test_utils;
 
-pub(crate) use halo2::types::{CircuitInstance, MerklePath};
+pub(crate) use halo2::types::{CircuitInstance, CircuitMerkleTreeLeaf, CircuitWitness, MerklePath};
