+++ mithril-stm/src/proof_system/halo2_snark/mod.rs
@@ -2,9 +2,9 @@ mod aggregate_key;
 mod clerk;
 mod eligibility;
 mod message;
+mod prover_input;
 mod signer;
 mod single_signature;
-mod witness;
 
 pub use aggregate_key::AggregateVerificationKeyForSnark;
 pub(crate) use clerk::SnarkClerk;
