--- mithril-stm/src/proof_system/halo2_snark/witness/mod.rs
@@ -1,4 +1,0 @@
-mod prover_input;
-mod witness_entry;
-
-pub(super) use witness_entry::WitnessEntry;
